♻️ refactor: add middleware and state for ListView
- introduced middleware component for ListView data parsing and state management
- improved handling of data structure and rendering state within the middleware layer
