+++ src/components/Media/MediaSection/ListView/components/MainNewsItem.jsx
